This builds for me under the following configurations:

clang++-3.4-libc++-3.5-debug
clang++-3.4-libc++-3.5-release
clang++-3.4-libstdc++-4.9-debug
clang++-3.4-libstdc++-4.9-release
clang++-3.5-libc++-3.5-debug
clang++-3.5-libc++-3.5-release
clang++-3.5-libstdc++-4.9-debug
clang++-3.5-libstdc++-4.9-release
clang++-3.6-libc++-3.5-debug
clang++-3.6-libc++-3.5-release
clang++-3.6-libstdc++-4.9-debug
clang++-3.6-libstdc++-4.9-release
clang++-3.7-libc++-3.5-debug
clang++-3.7-libc++-3.5-release
clang++-3.7-libstdc++-4.9-debug
clang++-3.7-libstdc++-4.9-release
g++-4.4-debug
g++-4.4-release
g++-4.5-debug
g++-4.5-release
g++-4.6-debug
g++-4.6-release
g++-4.7-debug
g++-4.7-release
g++-4.8-debug
g++-4.8-release
g++-4.9-debug
g++-4.9-release
g++-5.1-debug
g++-5.1-release

Close #9: port to clang

+++ RPC/ClientSessionTest.cc
@@ -231,8 +231,9 @@ TEST_F(RPCClientSessionTest, constructor_timeout_TimingSensitive) {
         Core::Time::SteadyClock::now() + std::chrono::milliseconds(5),
         Core::Config());
     auto end = Core::Time::SystemClock::now();
-    uint64_t elapsedMs = std::chrono::duration_cast<std::chrono::milliseconds>(
-                            end - start).count();
+    uint64_t elapsedMs =
+        uint64_t(std::chrono::duration_cast<std::chrono::milliseconds>(
+                            end - start).count());
     EXPECT_LE(5U, elapsedMs);
     EXPECT_GE(100U, elapsedMs);
     ClientSession::connectFn = ::connect;
